pyproject.toml
[project]
name = "rest-api-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "django>=6.0.1",
  "django-filter>=25.2",
  "djangorestframework>=3.16.1",
  "djangorestframework-simplejwt>=5.5.1",
  "pillow>=12.1.0",
]

[dependency-groups]
dev = [
  "django-stubs>=5.2.8",
  "django-types>=0.22.0",
  "djangorestframework-stubs>=3.16.7",
  "faker>=40.1.2",
]


[tool.pyright]
reportIncompatibleVariableOverride = false
